fixed surf match distance int to double

+++ localization/depth_odometry/include/depth_odometry/surf_feature_detector_and_matcher_params.h
@@ -21,7 +21,7 @@
 namespace depth_odometry {
 struct SurfFeatureDetectorAndMatcherParams {
   int surf_threshold;
-  int max_match_distance;
+  double max_match_distance;
 };
 }  // namespace depth_odometry
